requirements.txt
torch
nibabel==3.2
dgl
matplotlib
medpy
python-dotenv